requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.5
langchain-community>=0.2.0
langchain-openai>=0.1.0
pypdf>=4.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
openai>=1.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
